fix: don't keep updating because of defaulted annotations
The Cloudflare provider supports a special annotation to tell
Cloudflare whether or not to proxy for the hostname in question.  It
also has a default setting that it will use if the annotation isn't
specified.  Unfortunately, because it always returns the value read
from Cloudflare, if you don't explicitly specify it on your ingresses,
External DNS will update the DNS records over and over because the
ingresses' lists of provider specific attributes don't match the ones
the provider claims it has.

The fix for this is to allow the provider to intervene in the
comparison process, so that it can tell the planner that e.g. the lack
of a particular annotation is the same as setting that thing to true
or false.

Fixes issue #1463.